+++ NEWS.md
@@ -1,3 +1,9 @@
+# ospsuite.utils 1.4.0
+
+NEW FUNCTIONS
+
+* `ifEqual()` and `ifIncluded()` for conditional values.
+
 # ospsuite.utils 1.3.0
 
 NEW FUNCTIONS